Refactor editor modes into instances
An editor mode had information in models/EditorMode and widgets/CommandBar
which should be bundled in a single place. In this way it is easier to
see all information regarding a EditorMode and change/add new ones.

EditorMode has the following attributes: `id`, `name`, `description`, `text`;
which tries to capture all information that was present in the old
implementation. `id` might be superfluous if the numerical property is
nowhere used. `name` is doubled from the variable name that holds the
instance. It violates the DRY principle for an easier implementation.
`description` captures the information that was present in a comment in
the old models/EditorMode.py . `text` was present in
`widgets/CommandBar`.

Instead of class attributes now module attributes are used.
`models/__init__.py` had to be changed to reflect that change.